pyproject.toml
[project]
name = "agmlib"
version = "0.2.0"
description = "AGM Training library scaffold: entity-first, event-driven adaptive RL controllers"
authors = [{ name = "Datamutant.ai" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.6",
    "PyYAML>=6.0",
    "numpy>=1.24",
    "typing-extensions>=4.8",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"


